requirements.txt
Django==2.1.4
django-js-asset==1.1.0
django-mptt==0.9.1
pyenv==0.1
pytz==2018.7
